update format_state for polars DF
A polars DF return a list of columns by default and does not have a .values method. The Pandas DF .values api is a shortcut to return a numpy array anyway so I used that.
This is just what I do to make polars work in my personal project but should probably be tested better.