state: track ComputedVar dependencies
When a ComputedVar function accesses a Var, ComputedVar, or backend var via a
state instance, the `computed_var_dependencies` dict will keep track of
`requested_variable -> Set[dependent ComputedVar]`. This is implemented via
introspecting the stack to find the previous frame (__getattribute__ of the
ComputedVar).

During `get_delta` calculation, instead of recalculating all ComputedVar, only
recalculate the ComputedVar that depend on the set of changed vars. This
determination is made via recursive function to accomodate ComputedVar that
depend on ComputedVar.