vars: Add ComputedVars.deps function to calculate state Var dependencies
The `deps` function uses the stdlib `dis` module to iterate through the
instructions of the ComputedVar getter function. It keeps track of when
attributes are loaded from "self" and marks those as dependencies. If the
function attempts to call a method of the State, then recursively call `deps`
against the method and update the dependency set with the method's Var
depndencies.

Potentially addresses #870