state: track computed_var_dependency for dynamic route vars
When setting up dynamic route variables, ensure the ComputedVar has a reference
to the State class that it is attached to, for dependency tracking purposes.

Additionally set the accessor function name to match `param`, so that access
the dynamic var via `State.my_dynamic_var` gets the correct name (instead of
inner_func).