hydrate_middleware: ensure initial state update is_hydrated=False
As of #965, the the client side tells the backend that hydrate is complete.
When the browser requests a rehydrate, it will get a full copy of the current
state, which now includes is_hydrated=True. Then the `on_load` handlers will
run, which could take some time, then the client tells the server to
set_is_hydrated(True), even though it already is set True.

Resetting is_hydrated=False in the initial delta signals to the client that
additional on_load events may be needed before the state is considered
hydrated. This allows the frontend to display loading indicators based on
`state.is_hydrated` until all on_load handlers have returned.