test_create_config_e2e: end to end test of pcconfig creation
Use prerequisites.create_config to create an actual config file.

Then exec the contents of the file in the current python interpreter
to make sure it is valid code.

Finally, make assertions about the expected contents of the config
as python objects; specifically app_name, db_url, and env must be
set as expected.

Regression test for #1021