test_state: add_var with backend var
does not seem like this should work with the current implementation, but
indeed, it seems to.

i actually tried to write the code in state.py _propagate_var to put backend
vars in the `backend_vars` dict, but that actually didn't work...

so leaving the tests here, since they do pass, and hopefully further
refactoring might dislodge something in this area that can be improved.