Expose NoSSRComponent for downstream use
Previously, it wasn't really possible to access NoSSRComponent without doing a
deep import, which is discouraged since it makes future refactoring harder.

Exposing a commonly used Component base (especially when using complex third
party components that interact with media and Web APIs) makes it easier for the
community to wrap a wider array of React components.