test_config: fix `get_config` mocking
The test was mocking `reflex.config.get_config`, but it was calling the already-imported name `get_config`, which was NOT mocked. So these tests were NOT actually testing the passed `kwargs`.

Import reflex.config and deref `get_config` from the module to get the patched version.

Ensure that the config returned from the patched function is what we expect it to be.

Add additional test cases for `_event`-based event_namespace attributes.