useEventLoop with useReducer extracted from page template
REF-169: Move event loop logic out of the page template and define it in
state.js instead. This allows the templated logic to be simplified and
allow modifications to the event loop easier to reason about.

REF-170: Replace useState for state and result with useReducer.
Simplify logic and reduce unnecessary re-rendering in the frontend
by maintaining the event queue outside of the reactive state. Since the
event queue is never used for frontend rendering,
it doesn't really make sense to store it in reactive state.
With this change, on event re-renders are reduced by 75%.

Bonus: only include route change logic for dynamic pages. Otherwise the
hydrate event is sent twice when navigating client-side with NextLink.
This is still an issue for dynamic pages on the first load, but there's
no easy way around that with the current architecture.
When the state is pulled up to the App level with useContext, then
this can be fully addressed for all page types.