format_dict: only unescape quotes when removing outer quotes
If the outer quotes are removed, then it's safe and necessary to remove the
escapes on the inner quotes. Otherwise they should be left alone or they will
break the double-quote terminated string.

Only remove double quote escapes if we are for sure removing outer quotes with
a curly brace; a string containing the pattern `}"` will _never_ match this
regex because the quote following the curly will always be escaped by the JSON
encoder. So our regex here will only ever match a curly followed by JSON
encoder's closing double quote (which is always unescaped).