Client-side Storage API
Extend State to allow the storage of select Vars in the user's browser.

At compile time, extract all `rx.Cookie` and `rx.LocalStorage` typed state Vars
and save them in `/utils/context.js` as `clientStorage` for access on the
frontend.

When `rx.Cookie` or `rx.LocalStorage` values are set on the backend, the
frontend will also set them in the appropriate browser store based on the delta
received over the websocket. The clientStorage object provides options and
other information about how to persist the vars.

During initial hydrate event, values in the browser store will be read on the
frontend and sent to the backend as part of the event payload to be applied to
the state with HydrateMiddleware.

REF-172