Avoid compile step when running production backend
When executing `reflex run --env prod`, set `REFLEX_SKIP_COMPILE=yes` and then
during `App.compile`, exit early if this env var is set. This avoids running
the compile needlessly, since the prod backend does NOT use the compiled output
for anything.

In DEV mode, we must keep the compile step, because uvicorn re-importing the
app module and triggering the compile on the backend is what triggers the
NextJS frontend to rebuild.