Process EventHandler with is_background property set.
When encountering a background task EventHandler, move processing to an asyncio
task and yield a final state update to unblock other client events.

Background tasks do not get direct access to the State, but instead get `self`
as a `StateProxy`, which protects against mutation of the vars. The
`StateProxy` exposes async contextmanager protocol which refreshes the internal
proxied instance from the state_manager and holds an exclusive lock on the
token state while in context, allowing safe, stable reads and writes of state
vars from a long running background process.