Var._decode() gets the python value of a given Var
Calling an EventHandler returns an EventSpec with the function call arguments
encoded as `Var`. This works fine when formatting the Var into an event via
`format.format_event`, however when directly creating a payload and serializing
the entire Event to send to the client to be queued, the `fix_events` function
was just taking the "name" of the var, which is a json-encoded value. In order
to properly serialize the event payload, these Var args need to be decoded back
into their python values, so that the payload as a whole can be re-serialized.

Implement a `Var._decode` operation, which operates (mostly) as the inverse of
`Var.create`. This is an underscore method, because it actually doesn't return
another Var, but instead the real python value that the Var is created from.

Fix #1722