remove unnecessary setRouteNotFound(false) calls
we only need CSR once, when the user loads a URL and isn't already in the app;
once we're actually in the app, then we will never hit this Effect again.

even when navigating to a missing page, since the next router won't know about
it, it will trigger normal 404 behavior, reload the app and dump back into
setRouteNotFound(true).