Stream Upload StateUpdate objects over HTTP REST
The upload handler may yield multiple deltas and events, and these should not
wait until the entire upload handler completes, since it may take some time to
send the uploaded files to disk/s3 and it would be nice to send updates to the
frontend.

Since axios does not support streaming output natively in the browser, we reuse
the onDownloadProgress functionality to get access to the partial response,
split the JSON structures, and hand them off to the websocket.on("event")
handler, to be processed as-if they came from the websocket itself.

applyRestEvent no longer waits for the upload to complete, and all events from
the backend upload handler are marked as final so that upload events and
chained events do not block the websocket event queue.