fix test_state_with_invalid_yield now that Event _is_ a valid yield
This test was actually checking the wrong thing before and was just failing by
happenstance of passing an `Event` to `_check_valid`... it wasn't actually
calling the `invalid_handler` at any point.

So now the test actually calls `_process` with the handler, and checks for the
window alert and printed TypeError message on stdout.