test_app: remove all type: ignore comments on `app.process`
These were here because app.pyi was incorrectly marking process as `async` and
also giving its return type as `AsyncIterator`, but since the body did not have
a `yield` statement (due to being a stub), the type checker didn't realize it
was actually an iterator.

Removing the `async` keyword from the stub allows type checking to pass for
this function.