encode _var_state in formatted var names
Otherwise the resulting string when using f-string loses the state its
associated with and cannot be used to auto-memoization.

In every context that matters to the compiler, the f-string will be converted
_back_ into a Var, which reverses this process.

Decode _var_state in __post_init__ so that it always happens

Regardless of how the var is instantiated, we should never be including the var
markup in the final _var_name.

Part of REF-564