Var.foreach: fix index and iter type
Inspect the foreach function and pass the correct
arguments based on how many arguments it takes.

Use a unique index var name so that Var.foreach can
be nested inside each other.

Pass the correct _var_type for the arg -- it should
be the inner type of the container, not the container
itself.