Register _encode_var as serializer for Var type
Whenever Var appears within some other structure being JSON-ified, convert it with _encode_var to retain the _var_data.

Since `serialize` is only called during Var initialization, and then
immediately afterwards the _var_name is decoded to extract the VarData, this is
a safe operation. When the Var is encoded and then the entire structure is
encoded as JSON, _decode_var should try to double-decode VarData on failure.