UploadFilesProvider / UploadFilesContext: app tracks upload files
Now that any component can be pulled out of the page, it's not feasible to
share useState values directly via a global in state.js, because the components
accessing the global cannot update when the state changes (because React has no
way of tracking this).

Instead, create an AppWrap context that holds the selected upload files for the
entire app and have each component `useContext` to access it, such that it can
be re-rendered when selected files state is updated.

This requires a kind of gross special case to handle clearing of selected files
via call_script, so that it can be done from backend or frontend. A function to
clear selected files for a given ID is saved into the global "refs" object with
a double underscore name.