code_block: always pass code via "children" prop
The SyntaxHighlighter component is more reliable when the code itself is passed
via prop, rather than passed as an actual child. So save any passed child to
the `code` prop. And then render the contents of the `code` prop as `children=`
special prop (similar to how the rx.markdown component did it)

Using this new mechanism, the SyntaxHighlighter can work with rx.markdown at
arbitrary nesting levels (for example, when used with a custom component_map)