MemoizationMode: an extensible field of Component
To help handle memoization exceptions and special cases as they are discovered,
use a field on Component which controls aspecs of the auto-memoization process.

For now, only the "recursive" boolean is understood.

There is certain to be future categories of exceptions to memoization that will
be discovered, so instead of plumbing additional fields through, the
MemoizationMode can be extended with defaults and used to handle these special
cases in the components that need them, rather than centrally in the Component
or StatefulComponent logic.

Replace special case recursive logic for Foreach and InputGroup components with
this new mechanism.