Track changes applied to `Base` subclass via helper method.
When a helper method manipulates fields on a Base instance that is associated
with a state, recursively wrap the helper method with the MutableProxy so it
also tracks changes.

If the helper method returns a mutable object, this object will also be tied
back to the same field on the state to allow for "get_*" type functions that
still track changes.

Fix #2239