Radix themes: replace `color_scheme` with `color` in get_fields
This change ensures that the components never emit an actual `color_scheme`
prop to the generated code. Instead of treating the `color` prop swap in
`create` as a CSS prop, it will instead will emit `color={...}` as a direct
prop. Handling `color` as a style prop already occurs explicitly in `create`.

REF-1815