Move `is_used` to Upload component rather than UploadFilesProvider
The UploadFilesProvider is only created when actually compiling the frontend,
however the _upload endpoint needs to be added on the backend. The backend
skips the compile step, so it cannot detect whether UploadFilesProvider
actually gets used.

Moving the `is_used` to the `Upload` component itself means that just importing
an app which uses Upload can be detected without compiling anything.