[REF-2086] Avoid "Warning: The path to the Node binary could not be found."
When bootstrapping node itself, we do not need to warn the user that node
cannot be found -- it's obviously not found because we're installing it right
now. This warning leads some folks to believe that they need to install node
themselves (for example in a docker container), when in reality, reflex will
always try to install it's own known-good version of node first.

Also avoid a nasty traceback when Reflex attempts to run node or npm, but it's
not actually installed (instead, the warning will be displayed and it will say
"Invalid Command").

Move the recently added check for node before installing node above the
conditional so it can also skip node installation on Windows, which is
especially slow.