[REF-2392] Expose next.config.js transpilePackages key
Add `transpile_packages` knob to reflex Config object to escape hatch for
packages that need to be transpiled first.

Why not include this in the ImportVar mechanism? Because it's very nextjs
specific, and not needed in the general case, so I didn't want to plumb it
through the internals of the reflex Component system.

Fix #2933