Add some minimal validation of pyproject.toml
Currently the reflex component publish command only does some
validation of the pyproject.toml after it has extracted the
project name and version, and only if the user selects yes to
review interactively. As a result, a KeyError is raised if
the pyproject.toml has no project section, name, or version
defined.

This change adds a validation function to ensure those
projects are defined at a minimum and informs the user if that
is the case.

Closes #3338

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>