[ENG-5696] unwrap Var-typed annotations in @rx.memo signature (#5223)
Annotating the function arguments as `Var` might lead to more accurate type
hinting, but this breaks because the annotation was being unconditionally
wrapped in a Var.

With this change, if the annotation is already a Var, then the inner type is
used when constructing the dummy var that will be passed to the function for
evaluation.

Fix #5217