ENG-6500: Add vite-plugin-safari-cachebust to smooth out HMR issues (#5513)
* ENG-6500: Add vite-plugin-safari-cachebust to smooth out HMR issues

Safari with react-router has a fairly long standing issue where key modules are
aggressively cached, which results in code mismatch during hot reload that
cannot be resolved even by refreshing the browser. It requires a Shift+Refresh
to bypass the cache, which isn't a great DX.

This plugin rewrites the resulting HTML in dev mode to add timestamp query
parameter to modulepreload links in an attempt to get Safari to reload stuff
from the network. Not needed in prod mode because those files already contain a
content hash in the filename.

* Simplify logic of rewriteModuleImports

Generate an array of pattern, replacement pairs, and then reduce those over the
original html at the end.

* Convert plugin to JS with jsdoc

* console log when the plugin is first activated

* rephrase log message

* rich markup escape streamed process output