ENG-6503: use react-router Link instead of rx.el.a (#5529)
* ENG-6503: use react-router Link instead of rx.el.a

React Router's Link is a thin wrapper over the plain A element, with
client-side routing support. It internally detects absolute links (even those
given by a Var at runtime), so there's no good reason why it shouldn't be used.

If downstream users really want the plain link, they can get it through
`rx.el.elements.a` or `rx.el.A.create`, but the use case should be quite rare.

Updates to lazy_loader to allow overriding the returned `__all__` with extra
imports that are not in the submodules or submod_attrs.

* Lazily resolve imports to avoid circular dependency

* Update test expectations "a" -> ReactRouterLink

* Update ReactRouterLink _invalid_children

* update pyi_hashes.json