ENG-7314: Only check for valid app name once per config instance (#5710)
* ENG-7314: Only check for valid app name once per config instance

After the backend comes up, we check the app name once and validate that it's
good. There is no need to check it again in the same process.

Fixes the spurious `app.app module not found` errors that occur when deleting
and replacing the main app module while the backend is running, but before
triggering a reload.

* Update reflex/config.py

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>