Automatic websocket reconnect and reload handling (#5805)
* emit_update: take `token` instead of `sid`

This allows the app to be more resilient in the face of websocket reconnects.
The event is processed against a token, so there's no reason to maintain
websocket affinity for event processing. Whenever the update is ready to send,
it will be sent to the current websocket/sid associated.

* Automatic websocket reconnect and reload handling

* ensureSocketConnected is called when adding events or pumping the queue to
  trigger an automatic reconnection to the backend
* when "reload" event is encountered, trigger a re-hydrate and wait until ALL
  on_load have finished processing and `is_hydrated` is True before requeue the
  event that caused the "reload"

* Update mock token_to_sid mapping for test

* Add disconnect/reconnect test to test_background_task

* Remove non-background disconnect/reconnect test

It doesn't really work, because the frontend will only process one
non-background event at a time, so the disconnect ends up occuring after the
event handler is already done.